Simplify API by folding together channels and subscriptions
ActionCable doesn't allow you to perform any actions on channels you're not subscribed to, so there's not much value of a channel object that hasn't been subscribed, and it's pretty common to want to wait for the channel subscription to be confirmed, which necessitated using the statuses stream in the old factoring. This was unnecessary complexity.